Roles and DB initialization

+++ backend/controllers/article.controller.js
@@ -1,18 +1,14 @@
 const db = require("../models");
-const Article = db.articles;
-const User = db.users;
+const Article = db.article;
+const User = db.user;
 const fs = require("fs");
-const Op = db.Sequelize.Op;
-
-const jwt = require("jsonwebtoken");
-const userModel = require("../models/user.model");
 
 exports.create = (req, res) => {
   const title = req.body.title;
   const description = req.body.description;
-  const image = `${req.protocol}://${req.get("host")}/images/${
+  const image = req.file ? `${req.protocol}://${req.get("host")}/images/${
     req.file.filename
-  }`;
+  }` : null;
 
   if (title == "" || description == "") {
     return res.status(400).send({
@@ -70,22 +66,35 @@ exports.findOne = (req, res) => {
     });
 };
 
-exports.update = (req, res) => {};
+exports.update = (req, res) => {
+
+};
 
 exports.trash = (req, res) => {
   const myId = req.params.id;
 
   Article.findByPk(myId)
     .then((data) => {
-      const filename = data.image.split("/images/")[1];
-      fs.unlink(`images/${filename}`, () => {
+      if (data.image) {
+        const filename = data.image.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Article.destroy({
+            where: {
+              id: myId,
+            },
+          })
+            .then(res.status(200).send({ message: "Article delete!" }))
+        });
+      } else {
+        
         Article.destroy({
           where: {
             id: myId,
           },
         })
           .then(res.status(200).send({ message: "Article delete!" }))
-      });
+      }
+
     })
     .catch((error) => res.status(500).json({ error }));
 };
